service/scheduler: suspend, extract the suspend code to a separate file
